+++ example/pubspec.yaml
@@ -8,9 +8,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  path_provider: ^1.3.0
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
 
 dev_dependencies:
@@ -20,16 +17,8 @@ dev_dependencies:
   jisho:
     path: ../
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-  # To add assets to your application, add an assets section, like this:
   assets:
     - packages/jisho/index/080d0121166f4c8e98bf4fb0d90b272f.fieldnorm
     - packages/jisho/index/.tantivy-meta.lock
@@ -56,26 +45,3 @@ flutter:
     - packages/jisho/index/080d0121166f4c8e98bf4fb0d90b272f.idx
     - packages/jisho/index/080d0121166f4c8e98bf4fb0d90b272f.fast
     - packages/jisho/index/080d0121166f4c8e98bf4fb0d90b272f.store
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev-and-images/#resolution-aware.
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev-and-images/#from-packages
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
