+++ lib/src/types.dart
@@ -3,13 +3,3 @@ import 'package:ffi/ffi.dart';
 
 typedef NativeRustSearchFunction = ffi.Pointer<Utf8> Function(
     ffi.Pointer<Utf8>, ffi.Pointer<Utf8>);
-
-typedef NativeSearchFunction = ffi.Pointer<Utf8> Function(
-    ffi.Pointer<Utf8>, ffi.Pointer<Utf8>);
-
-typedef NativeRustSearchImFunction = ffi.Pointer<Utf8> Function(
-    ffi.Pointer<Utf8>);
-
-typedef NativeRustReadfileFunction = ffi.Pointer<Utf8> Function(
-    ffi.Pointer<Utf8>);
-typedef NativeSearchImFunction = ffi.Pointer<Utf8> Function(ffi.Pointer<Utf8>);
