+++ pubspec.yaml
@@ -11,24 +11,16 @@ dependencies:
   flutter:
     sdk: flutter
   ffi: ^0.1.1
+  path_provider: ^1.3.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-  # This section identifies this Flutter project as a plugin project.
-  # The androidPackage and pluginClass identifiers should not ordinarily
-  # be modified. They are used by the tooling to maintain consistency when
-  # adding or updating assets for this project.
   plugin:
     androidPackage: com.example.jisho
     pluginClass: JishoPlugin
-  # To add assets to your plugin package, add an assets section, like this:
   assets:
     - lib/index/080d0121166f4c8e98bf4fb0d90b272f.fieldnorm
     - lib/index/.tantivy-meta.lock
@@ -55,28 +47,3 @@ flutter:
     - lib/index/080d0121166f4c8e98bf4fb0d90b272f.idx
     - lib/index/080d0121166f4c8e98bf4fb0d90b272f.fast
     - lib/index/080d0121166f4c8e98bf4fb0d90b272f.store
-  #
-  # For details regarding assets in packages, see
-  # https://flutter.dev-and-images/#from-packages
-  #
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev-and-images/#resolution-aware.
-  # To add custom fonts to your plugin package, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts in packages, see
-  # https://flutter.dev/custom-fonts/#from-packages
